ai-service/requirements.txt
# --- CORE DEPENDENCIES ---
Flask==2.3.0
flask-cors==4.0.0
pymongo==4.6.0
PyPDF2==3.0.1
numpy==1.24.3
pdfplumber==0.10.3
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
python-docx==1.1.0
joblib==1.3.2

# --- LANGCHAIN STACK (COMPATIBLE SET) ---
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-google-genai==1.0.3

# --- GOOGLE GENAI ---
google-ai-generativelanguage==0.6.4
google-generativeai==0.5.4

# --- EMBEDDINGS / NLP ---
sentence-transformers==2.2.2

# --- AUDIO ---
pydub==0.25.1
soundfile==0.12.1

# --- LANGGRAPH ---
langgraph==0.0.44
